refactor(gcl): restructure hands-on guide routes and tests under app/gcl/hands-on
