+++ app/gcl/hands-on/build-a-secure-google-cloud-network/NavBar.tsx
