+++ app/gcl/hands-on/cloud-load-balancing-guide/NavBar.tsx
