+++ app/gcl/hands-on/develop-your-gcp-network/NavBar.tsx
