+++ app/gcl/hands-on/develop-your-gcp-network/constants.ts
