+++ app/gcl/hands-on/gcp-security-fundamentals-guide/NavBar.tsx
