+++ app/gcl/hands-on/gcp-security-fundamentals-guide/constants.ts
