+++ app/gcl/hands-on/set-up-an-app-dev-environment-on-google-cloud/constants.ts
